Add test, for 201 response on post

+++ scripts/create_release.sh
@@ -1,6 +1,7 @@
 #!/usr/bin/env bash
 
 set -Eeou pipefail
+# Create a github release, on master
 
 # Ensure we get the correct param count
 export USAGE="Usage: $0 <env> <tag> <release title> <release desc> <github token>  e.g. ${0} uat \"v1.1.0\" \"my tag title\" \"env=uat add new stuff\" abcefdummytokenghijk"
@@ -24,6 +25,7 @@ then
   printf "Env must be either uat or prod\n" && exit 1
 fi
 
+export TARGET_BRANCH=master
 export GIT_BRANCH=$(git rev-parse --abbrev-ref HEAD | sed "s*/*-*g")
 export GIT_COMMIT=$(git rev-parse --short --verify HEAD)
 export GIT_COMMITER=$(git show -s --format='%ae' "${GIT_COMMIT}")
@@ -31,6 +33,10 @@ export GIT_OWNER=$(git remote get-url origin | sed 's#.*github.com/##' | xargs d
 export GIT_REPO=$(git remote get-url origin | sed 's#.*github.com/##' | xargs basename | sed 's#.git##g')
 export CURRENT_USER=$(whoami | sed "s/[^[:alpha:]]//g")
 export RELEASE_DESCRIPTION="env=${TARGET_ENV} released-by=${CURRENT_USER} - [${RELEASE_DESC}]"
+
+if [[ "${TARGET_BRANCH}" != "${GIT_BRANCH}" ]]; then
+  printf "\n\nWarning: current branch, ${GIT_BRANCH}, is not ${TARGET_BRANCH}, tag will be created on ${TARGET_BRANCH}\n\n"
+fi
 
 export SCHEME="https"
 export HOST="api.github.com"
@@ -50,7 +56,7 @@ TMP_FILE=./$$.tmp
 cat <<- PAYLOAD > "${TMP_FILE}"
 {
   "tag_name": "${RELEASE_TAG_NAME}",
-  "target_commitish": "master",
+  "target_commitish": "${TARGET_BRANCH}",
   "name": "${RELEASE_TITLE}",
   "body": "${RELEASE_DESCRIPTION}",
   "draft": false,
@@ -60,10 +66,22 @@ PAYLOAD
 
 jq . "${TMP_FILE}"
 
+export RESPONSE=$(
 curl -s -X POST \
   -H "Content-Type: application/json" \
   -H "Authorization: token ${GITHUB_TOKEN}" \
   -d "@${TMP_FILE}" \
-  "${RELEASES_ENDPOINT}"
+  "${RELEASES_ENDPOINT}" \
+  -w "\n%{http_code}")
+
+echo "RESPONSE=${RESPONSE}"
+
+# Ensure we got a 201, created, as per the spec
+if [[ "$(tail -n 1 <<< "${RESPONSE}")" != "201" ]];
+then
+  echo "Error ocurred"
+  exit 1
+fi
+
 
 rm -rf "${TMP_FILE}"
